+++ src/include/libdnn/profile.h
@@ -18,19 +18,19 @@
 	#define inc_mac_vec(v) (void)0
 	#define print_stats() (void)0
 #else
-void inc_ld(uint);
-void inc_st(uint);
-void inc_add(uint);
-void inc_mul(uint);
-void inc_addr_add(uint);
-void inc_addr_mul(uint);
+void inc_ld(uint16_t);
+void inc_st(uint16_t);
+void inc_add(uint16_t);
+void inc_mul(uint16_t);
+void inc_addr_add(uint16_t);
+void inc_addr_mul(uint16_t);
 
 	#ifdef CONFIG_LEA
-		void inc_ld_vec(uint);
-		void inc_st_vec(uint);
-		void inc_mac_vec(uint);
-		void inc_fir_vec(uint, uint);
-		void inc_add_vec(uint);
+		void inc_ld_vec(uint16_t);
+		void inc_st_vec(uint16_t);
+		void inc_mac_vec(uint16_t);
+		void inc_fir_vec(uint16_t, uint16_t);
+		void inc_add_vec(uint16_t);
 	#endif
 
 void print_stats();
