+++ src/profile.c
@@ -5,7 +5,7 @@
 
 #ifdef CONFIG_PROFILE
 typedef struct {
-	uint count;
+	uint16_t count;
 	unsigned long size;
 } vec_count;
 
@@ -21,51 +21,51 @@ static __fram vec_count add_vec_count = {.count = 0, .size = 0.};
 static __fram vec_count fir_vec_count = {.count = 0, .size = 0.};
 static __fram vec_count mac_vec_count = {.count = 0, .size = 0.};
 
-void inc_ld(uint v) {
+void inc_ld(uint16_t v) {
 	ld_count += v;
 }
 
-void inc_st(uint v) {
+void inc_st(uint16_t v) {
 	st_count += v;
 }
 
-void inc_ld_vec(uint v) {
+void inc_ld_vec(uint16_t v) {
 	ld_vec_count.size += v;
 	ld_vec_count.count++;
 }
 
-void inc_st_vec(uint v) {
+void inc_st_vec(uint16_t v) {
 	st_vec_count.size += v;
 	st_vec_count.count++;
 }
 
-void inc_add(uint v) {
+void inc_add(uint16_t v) {
 	add_count += v;
 }
 
-void inc_mul(uint v) {
+void inc_mul(uint16_t v) {
 	mul_count += v;
 }
 
-void inc_addr_add(uint v) {
+void inc_addr_add(uint16_t v) {
 	addr_add_count += v;
 }
 
-void inc_addr_mul(uint v) {
+void inc_addr_mul(uint16_t v) {
 	addr_mul_count += v;
 }
 
-void inc_mac_vec(uint v) {
+void inc_mac_vec(uint16_t v) {
 	mac_vec_count.size += v;
 	mac_vec_count.count++;
 }
 
-void inc_fir_vec(uint v, uint f) {
+void inc_fir_vec(uint16_t v, uint16_t f) {
 	fir_vec_count.size += v;
 	fir_vec_count.count++;
 }
 
-void inc_add_vec(uint v) {
+void inc_add_vec(uint16_t v) {
 	add_vec_count.size += v;
 	add_vec_count.count++;
 }
